requirements.txt
Django==4.2.7
psycopg2-binary==2.9.9
djangorestframework
markdown
django-filter
Pillow>=10